467 version 0032 updates (#476)
* Version 0.0.30 updates (#457) (#458)

* 20241125 @Mookse
- version

* 20241126 @Mookse
- avatar routine
- routine updates

* 20241126 @Mookse
- Rename Collection Bars #409

* 20241126 @Mookse
- Experience Agent #415
- experiences()

* 20241129 @Mookse
- wip unstable

* 20241201 @Mookse
- wip unstable

* 20241202 @Mookse
- wip unstable

* 20241202 @Mookse
- wip stable

* 20250207 @Mookse
- initial speech recognition in `guests.mjs`

* 20250215 @Mookse
- Speech-to-Text #454

* 20250216 @Mookse
- draft complete Speech-to-Text #454

* 20250216 @Mookse
- #454 console cleanup

* 20250217 @Mookse
- #456 Guest

* 20240217 @Mookse
- Global Chat #456 HTML

* 20250217 @Mookse
- Global Chat #456 Member

* 20250312 @Mookse
- version update

* 20250312 @Mookse
- cleaned Factory

* Initial style tweaksfor flex wrapping

* Add needed div to layout

* Fix bg image filepath

* 20250313 @Mookse
- minor adjustments: min screen width 320px for phones
- slight padding reduction
- returned translucence
- bkg color root defined

* 20250313 @Mookse
- Bot containers were growing unneccessarily,

* 20250313 @Mookse
- removed label class from audio-mic on members

* 20250313 @Mookse
- Deprecate MyLifeSession #413
- tested implementation, looks good
- alerts may stll need assistance

* 20250313 @Mookse
- Migrate `core.mjs` #414
- tested implementation

* 20250313 @Mookse
- Migrate `core.mjs` #414
- checkalerts() fix

* 20250312 @Mookse
- MyLife Member Services Platform Version 0.0.32 #467
- FIX: signup not switching div

* 202503113 @Mookse
- MyLife Member Services Platform Version 0.0.32 #467
- FIX: frontend alerts running twice

* 20250313 @Mookse
- MyLife Member Services Platform Version 0.0.32 #467
- removed core.mjs (already deprecated)

* 20250313 @Mookse
- center align page-content
- removed media query

* 20250314 @Mookse
- update about with new director content

* 20250314 @Mookse
- MyLife General Functioneer: make action callbacks #474
- Obscure()

* 20250314 @Mookse
- Functioneer Callbacks #474
- prepareSummary()

* 20250314 @Mookse
- FIX: direct updates to collection item summary

---------
