+++ inc/js/deprecated/core.mjs
@@ -1,308 +1,435 @@
-//	question/answer functions
-class Deprecated{
-    constructor(){}
-    //  public functions from class Member (core.mjs)
-    async assignPrimingQuestions(_question){
-        return this.buildFewShotQuestions(
-            await this.fetchEnquiryMetadata(_question)	//	what question type category is this?
-        )
-    }
-    buildFewShotQuestions(_category){
-        const _fewShotQuestions = []
-        switch(_category){
-            case 'abilities':	//	abilities & skills
-                _fewShotQuestions.push(
-                    this.chatObjectify(`${ this.memberName }'s abilities?`,false),
-                    this.chatObjectify(this.abilities),
-                    this.chatObjectify(`${ this.memberName }'s skills?`,false),
-                    this.chatObjectify(this.skills),
-                )
-                break
-            case 'artifacts':	//	artifacts & possessions
-            case 'beliefs':	//	beliefs & values
-                _fewShotQuestions.push(
-                    this.chatObjectify(`${ this.memberName }'s beliefs?`,false),
-                    this.chatObjectify(this.beliefs),
-                    this.chatObjectify(`${ this.memberName }'s values?`,false),
-                    this.chatObjectify(this.values),
-                )
-                break
-            case 'facts':	//	biological and historical facts
-                _fewShotQuestions.push(
-                    this.chatObjectify(`Some of ${ this.memberName }'s biological facts?`,false),
-                    this.chatObjectify(this.facts.biological),
-                    this.chatObjectify(`Some of ${ this.memberName }'s historical facts?`,false),
-                    this.chatObjectify(this.facts.historical),
-                )
-                break
-            case 'interests':	//	interests & hobbies
-                _fewShotQuestions.push(
-                    this.chatObjectify(`${ this.memberName }'s interests?`,false),
-                    this.chatObjectify(this.interests),
-                    this.chatObjectify(`${ this.memberName }'s hobbies?`,false),
-                    this.chatObjectify(this.hobbies),
-                )
-                break
-            case 'preferences':	//	preferences & beliefs
-                _fewShotQuestions.push(
-                    this.chatObjectify(`${ this.memberName }'s preferences?`,false),
-                    this.chatObjectify(this.preferences),
-                    this.chatObjectify(`${ this.memberName }'s beliefs?`,false),
-                    this.chatObjectify(this.beliefs),
-                )
-                break
-            case 'relations':
-            case 'other':
-            default:	//	motivations & beliefs
-                _fewShotQuestions.push(
-                    this.chatObjectify(`${ this.memberName }'s motivations?`,false),
-                    this.chatObjectify(this.motivations),
-                    this.chatObjectify(`${ this.memberName }'s beliefs?`,false),
-                    this.chatObjectify(this.beliefs),
-                )
-                break
-        }
-        return _fewShotQuestions
-    }
-    chatObjectify(_content,_bAgent=true){
-        return {
-            role: (_bAgent)?'assistant':'user',
-            content: _content
-        }
-    }
-    async encodeQuestion(_question){
-        const _model = 'text-davinci-001'
-        const _youReference = await this.personality.createCompletion({
-            model: _model,
-            prompt: `Is "you" in quote likely referring to ai-agent, human, or unknown?\nQuote: "${_question}"\nRefers to:`,
-            temperature: 0,
-            max_tokens: 60,
-            top_p: 1,
-            frequency_penalty: 0.5,
-            presence_penalty: 0,
-        })
-            .then(
-                (_response)=>{
-                    //	response insertion/alteration points for approval, validation, storage, pruning
-                    //	challengeResponse(_response) //	insertion point: human reviewable
-                    return _response.data.choices[0].text.trim().toLowerCase()
-                }
-            )
-            .catch(err=>{
-                console.log(err)
-                return 'unknown'	//	emit for server
-            })
-        //	youReference contains human
-        if(_youReference.includes('human')){
-            _question = _question.replace(/your/gi,`${this.memberName}'s`)	//	replace your with memberName's
-            _question = _question.replace(/you/gi,this.memberName)	//	replace you with memberName
-        }
-        return _question
-    } 
-    async fetchEnquiryMetadata(_question){	// human core
-        //	what is the best category for this question?
-        const _model = 'text-davinci-001'
-        const _category = await this.personality.createCompletion({
-            model: _model,
-            prompt: `What is the best category for this quote?\nCategories: ${ this.categories.toString() }\nQuote: \"${ _question }\"\nCategory:`,	//	user array of human categories
-            temperature: 0,
-            max_tokens: 60,
-            top_p: 1,
-            frequency_penalty: 0.5,
-            presence_penalty: 0,
-        })
-            .then(
-                (_response)=>{
-                    //	response insertion/alteration points for approval, validation, storage, pruning
-                    //	challengeResponse(_response) //	insertion point: human reviewable
-                    return _response.data.choices[0].text.trim().toLowerCase()
-                }
-            )
-            .catch(err=>{
-                console.log(err)
-                //	emit for server
-            })
-        const _categoryModeler = await this.dataservice.getItems('categorization','*')
-        if(_categoryModeler.length){	//	if storage easily accessible, use it
-            const _update = (_categoryModeler[0]?.[_model])
-                ?	[{ op: 'add', path: `/${ _model }/-`, value: { [_category]: _question } }]	//	add array value
-                :	[{ op: 'add', path: `/${ _model }`, value: [{ [_category]: _question }] }]	//	create array
-            this.dataservice.patchItem(	//	temporary log; move to different db and perform async
-    //	Error: PartitionKey extracted from document doesn't match the one specified in the header
-                _categoryModeler[0].id,
-                _update
-            )
-            console.log(chalk.bold.blueBright(_model,_category))
-        }
-        return _category
-    }
-    async formatQuestion(_question){
-        if(this.form==='human') _question = await this.encodeQuestion(_question)
-        return {
-            role: 'user',
-            content: _question
-        }
-    }
-    formatResponse(_str){
-        //	insert routines for emphasis
-        const _response=this.detokenize(_str.data.choices[0].message.content)
-            .replace(/(\s|^)mylife(\s|$)/gi, "$1<em>MyLife</em>$2")
-        return _response
-    }
-    //	misc functions
-    detokenize(_str){
-        return _str.replace(/<\|\|/g,'').replace(/\|\|>/g,'')
-    }
-	tokenize(_str){
-		return '<||'+_str+'||>'
+//	imports
+import EventEmitter from 'events'
+//	server-specific imports
+import initRouter from './routes.mjs'
+//	define export Classes for Members and MyLife
+class Member extends EventEmitter {
+	#avatar
+	#factory
+	constructor(factory){
+		super()
+		this.#factory = factory
 	}
-    //  public functions from class Organization (core.mjs)
-	async assignLocalContent(_question){	//	assign local content from pgvector
-		let _localContent = await this.dataservice.getLocalRecords(_question)
-		return (_localContent.length)	//	"micro"-prompt around content
-			?	[
-					{
-						role: 'user',
-						content: `What is top corporate record about: ${_question}`
-					},
-					{
-						role: 'assistant',
-						content: await this.dataservice.getLocalRecords(_question)
-					}
-				]
-			:	[]
+	/**
+	 * Initializes `this.#avatar` and returns `this`. The Avatar will thence be primary point of interaction with Session and Server. Only upon dissolution of Avatar [sessionEnd] is triggered an internal `this` evolution evaluation based on upon the conduct of Avatar.
+	 * @async
+	 * @public
+	 * @param {Q} avatar - Avatar object to initialize, when infused from child instance
+	 * @returns {Promise} Promise resolves to this Member class instantiation
+	 */
+	async init(avatar){
+		this.#avatar = avatar
+			?? await this.#factory.getAvatar()
+		return this
 	}
-	async assignPrimingQuestions(_question){	//	corporate version
-		return this.buildFewShotQuestions(
-			await this.fetchEnquiryType(_question)	//	what question type is this?
-		)
+	//	getter/setter functions
+	get abilities(){
+		return this.core.abilities
 	}
-	buildFewShotQuestions(_questionType){
-		//	cascade through, only apply other if no length currently [either failed in assigned case or was innately other]
-		const _fewShotQuestions = []
-		switch(_questionType){
-			case 'products': case 'services':	//	 services & roadmap 
-				_fewShotQuestions.push(
-					this.chatObjectify(`${ this.name }'s services?`,false),
-					this.chatObjectify(this.services),
-					this.chatObjectify(`${ this.name }'s technical roadmap?`,false),
-					this.chatObjectify(this.roadmap),
-				)
-				break
-			case 'customer': case 'support':	//	membership & services
-				_fewShotQuestions.push(
-					this.chatObjectify(`${ this.name }'s registration?`,false),
-					this.chatObjectify(this.membership),
-					this.chatObjectify(`${ this.name }'s services?`,false),
-					this.chatObjectify(this.services),
-				)
-				break
-			case 'security':	//	security & privacy
-				_fewShotQuestions.push(
-					this.chatObjectify(`${ this.name }'s security?`,false),
-					this.chatObjectify(this.security),
-					this.chatObjectify(`${ this.name }'s privacy policy?`,false),
-					this.chatObjectify(this.privacy),
-				)
-				break
-			case 'business': case 'info':	//	governance & vision
-				_fewShotQuestions.push(
-					this.chatObjectify(`${ this.name }'s governance?`,false),
-					this.chatObjectify(this.governance),
-					this.chatObjectify(`${ this.name }'s vision?`,false),
-					this.chatObjectify(this.vision),
-				)
-				break
-			case 'values':	//	values & philosophy
-				_fewShotQuestions.push(
-					this.chatObjectify(`${ this.name }'s values?`,false),
-					this.chatObjectify(this.values),
-					this.chatObjectify(`${ this.name }'s philosophy?`,false),
-					this.chatObjectify(this.philosophy),
-				)
-				break
-			case 'technology':	//	roadmap & security
-				_fewShotQuestions.push(
-					this.chatObjectify(`${ this.name }'s technical roadmap?`,false),
-					this.chatObjectify(this.roadmap),
-					this.chatObjectify(`${ this.name }'s security?`,false),
-					this.chatObjectify(this.security),
-				)
-				break
-			case 'other':
+	get agent(){
+		return this.#avatar
+	}
+	get agentCategories(){
+		return this.agent.categories
+	}
+	get agentCommand(){
+		return this.agent.command_word
+	}
+	get agentDescription(){	//	agent description (not required)
+		if(!this.agent?.description)
+			this.avatar.description = `I am ${ this.agentName }, AI-Agent for ${ this.name }`
+		return this.agent.description
+	}
+	get agentName(){
+		return this.avatar.names[0]
+	}
+	get agentProxy(){
+		switch(this.form){	//	need switch because I could not overload class function
+			case 'organization':
+				return this.description
 			default:
-				if(_fewShotQuestions.length <= 2){	//	if populated, requires user<->agent interaction
-					_fewShotQuestions.push(
-						this.chatObjectify(`${ this.name }'s mission?`,false),
-						this.chatObjectify(this.mission),
-						this.chatObjectify(`${ this.name }'s vision?`,false),
-						this.chatObjectify(this.vision),
-					)
-				}
+				return this.bio
 		}
-		return _fewShotQuestions
 	}
-	async fetchEnquiryType(_question){	//	categorize, log and return
-		const _model = 'text-babbage-001'
-		const _category = await this.personality.createCompletion({
-			model: _model,
-			prompt: `Give best category for Phrase about the nonprofit company MyLife.org\nCategories: ${ this.categories.toString() }\nPhrase: \"${ _question }\"\nCategory:`,
-			temperature: 0,
-			max_tokens: 32,
-			top_p: 1,
-			frequency_penalty: 0.5,
-			presence_penalty: 0,
+	get avatar(){
+		return this.#avatar
+	}
+	set avatar(_Avatar){
+		//	oops, hack around how to get dna of avatar class; review options [could block at factory-getter level, most efficient and logical]
+		if(!this.#factory.isAvatar(_Avatar))
+			throw new Error('avatar requires Avatar Class')
+		this.#avatar = _Avatar
+	}
+	/**
+	 * Gets being type. Members are _only_ human prototypes, and presents as human, even though the current manifestation is a datacore `core` being.
+	 * @returns {string} returns 'human'
+	*/
+	get being(){
+		return 'human'
+	}
+	get beliefs(){
+		return this.core.beliefs
+	}
+	get bio(){
+		return this.core.bio
+	}
+	get categories(){
+		return this.agent.categories
+	}
+	get chat(){
+		return this.agent.chat
+	}
+	get consent(){
+		return this.#factory.consent	//	**caution**: returns <<PROMISE>>
+	}
+	set consent(_consent){
+		this.#factory.consents.unshift(_consent.id)
+	}
+	get core(){
+		return this.#factory.core
+	}
+	get dataservice(){
+		return this.dataservices
+	}
+	get dataservices(){
+		return this.#factory.dataservices
+	}
+	get description(){
+		return this.core.description
+	}
+	get email(){
+		return this.core.email
+	}
+	get factory(){
+		return this.#factory
+	}
+	get facts(){
+		return this.core.facts
+	}
+	get form(){
+		return this.core.form
+	}
+	get globals(){
+		return this.#factory.globals
+	}
+	get hobbies(){
+		return this.core.hobbies
+	}
+	get interests(){
+		return this.core.interests
+	}
+	get id(){
+		return this.sysid
+	}
+	get mbr_id(){
+		return this.#factory.mbr_id
+	}
+	get member(){
+		return this.core
+	}
+	get memberName(){
+		return this.core.names[0]
+	}
+	get name(){
+		return this.core.name
+	}
+	get newGuid(){
+		return this.globals.newGuid
+	}
+	get preferences(){
+		return this.core.preferences
+	}
+	get schemas(){
+		return this.#factory.schemas
+	}
+	get skills(){
+		return this.core.skills
+	}
+	get sysname(){
+		return this.mbr_id.split('|')[0]
+	}	
+	get sysid(){
+		return this.mbr_id.split('|')[1]
+	}
+	get values(){
+		return this.core.values
+	}
+	//	public functions
+	async testEmitters(){
+		//	test emitters with callbacks
+		this.emit('testEmitter',_response=>{
+
 		})
-			.then(
-				(_response)=>{
-					//	response insertion/alteration points for approval, validation, storage, pruning
-					//	challengeResponse(_response) //	insertion point: human reviewable
-					return _response.data.choices[0].text.trim().toLowerCase()
+	}
+}
+class Organization extends Member {	//	form=organization
+	#Menu
+	#Router
+	constructor(Factory){
+		super(Factory)
+	}
+	/* public functions */
+	async init(avatar){
+		if(!avatar)
+			throw new Error('avatar parameter currently required for Organization imprint')
+		return await super.init(avatar)
+	}
+	/* getters/setters */
+	/**
+	 * Gets being type. MyLife are _only_ MyLife prototypes, and presents as mylife, even though the current manifestation is a datacore `core` being.
+	 * @returns {string} returns 'MyLife'
+	*/
+	get being(){
+		return 'MyLife'
+	}
+	get description(){
+		return this.core.description
+	}
+	get governance(){
+		return this.core.governance
+	}
+	get membership(){
+		return this.core.membership
+	}
+	get menu(){
+		if(!this.#Menu){
+			this.#Menu = new (this.schemas.menu)(this).menu
+		}
+		return this.#Menu
+	}
+	get mission(){
+		return this.core.mission
+	}
+	get name(){
+		return this.core.names[0]
+	}
+	get philosophy(){
+		return this.core.philosophy
+	}
+	get privacy(){
+		return this.core.privacy
+	}
+	get roadmap(){
+		return this.core.roadmap
+	}
+	get router(){
+		if(!this.#Router){
+			this.#Router = initRouter(new (this.schemas.menu)(this))
+		}
+		return this.#Router
+	}
+	get security(){
+		return this.core.security
+	}
+	get services(){
+		return this.core.services
+	}
+	get values(){
+		return this.core.values
+	}
+	get version(){
+		return this.core.version ?? '0.0.17'
+	}
+	get vision(){
+		return this.core.vision
+	}
+}
+class MyLife extends Organization {	// form=server
+	#avatar // MyLife's private class avatar, _same_ object reference as Member Class's `#avatar`
+	#currentShares = [] // array of current shares
+	#factory
+	#shareAgent
+	#version = '0.0.0' // indicates error
+	constructor(Factory){ // no session presumed to exist
+		super(Factory)
+		this.#factory = Factory
+	}
+	async init(){
+		this.#avatar = await this.#factory.getAvatar()
+		return await super.init(this.#avatar)
+	}
+	/* public functions */
+	/**
+	 * Retrieves all public experiences (i.e., owned by MyLife).
+	 * @returns {Object[]} - An array of the currently available public experiences.
+	 */
+	async availableExperiences(){
+		const experiences = ( await this.#factory.availableExperiences() )
+			.map(experience=>{ // map to display versions [from `mylife-avatar.mjs`]
+				const { autoplay=false, description, id, name, purpose, skippable=true,  } = experience
+				return {
+					description,
+					id,
+					name,
+					purpose,
 				}
-			)
-			.catch(err=>{
-				console.log(err)
-				return 'other'
-				//	emit for server
 			})
-		const _categoryModeler = await this.dataservice.getItems('categorization','*')
-		if(_categoryModeler.length){	//	if storage easily accessible, use it
-			const _update = (_categoryModeler[0]?.[_model])
-				?	[{ op: 'add', path: `/${ _model }/-`, value: { [_category]: _question } }]	//	add array value
-				:	[{ op: 'add', path: `/${ _model }`, value: [{ [_category]: _question }] }]	//	create array
-			this.dataservice.patchItem(	//	temporary log; move to different db and perform async
-	//	Error: PartitionKey extracted from document doesn't match the one specified in the header
-				_categoryModeler[0].id,
-				_update
-			)
-			console.log(chalk.bold.blueBright(_model,_category))
+		return experiences
+	}
+	/**
+	 * Challenges and logs in member.
+	 * @param {string} memberId - Member id to challenge.
+	 * @param {string} passphrase - Passphrase response to challenge.
+	 * @returns {boolean} - Whether or not member is logged in successfully.
+	 */
+	async challengeAccess(memberId, passphrase){
+		const challengeSuccessful = await this.#avatar.challengeAccess(memberId, passphrase)
+		return challengeSuccessful
+	}
+	/**
+	 * Returns the datacore object for the specified member id.
+	 * @param {string} mbr_id - The Member id to access datacore
+	 * @returns {Promise<object>} - Datacore object for member id
+	 */
+	async datacore(mbr_id){
+		if(!mbr_id || mbr_id===this.mbr_id)
+			throw new Error('datacore cannot be accessed')
+		const core = this.globals.sanitize(await this.#factory.datacore(mbr_id))
+		return core
+	}
+	/**
+	 * Submits and returns the journal or diary entry to MyLife via API.
+	 * @public
+	 * @todo - consent check-in with spawned Member Avatar
+	 * @param {object} summary - Object with story summary and metadata
+	 * @returns {object} - The story document from Cosmos.
+	 */
+	async entry(summary){
+		if(!summary.mbr_id?.length)
+			throw new Error('entry `mbr_id` required')
+		if(!summary.summary?.length)
+			throw new Error('entry `summary` required')
+		summary.being = 'entry'
+		summary.form = summary.form ?? 'journal'
+		return await this.summary(summary)
+	}
+	/**
+	 * Server MyLife _Maht instantiation uses this function to populate the most current alerts in the modular factory memoryspace. Currently only applicable to system types, but since this is implemented at the `core.mjs` scope, we can account
+	 * @public
+	 * @returns {void} returns nothing, performs operation
+	 */
+	getAlerts(){
+		this.#factory.getAlerts()
+	}
+	async getMyLifeSession(){
+		return await this.#factory.getMyLifeSession()
+	}
+	async hostedMemberList(){
+		let members = await this.hostedMembers()
+		return members.map(member=>member.mbr_id)
+	}
+	/**
+	 * Returns Array of hosted members based on validation requirements.
+	 * @param {Array} validations - Array of validation strings to filter membership.
+	 * @returns {Promise<Array>} - Array of string ids, one for each hosted member.
+	 */
+	async hostedMembers(validations){
+		return await this.#factory.hostedMembers(validations)
+	}
+	/**
+	 * Returns whether a specified member id is hosted on this instance.
+	 * @param {string} memberId - Member id
+	 * @returns {boolean} - Returns true if member is hosted
+	 */
+	async isMemberHosted(memberId){
+		const hostedMembers = await this.hostedMemberList()
+		const isHosted = hostedMembers.includes(memberId)
+		let isValidated = false
+        if(isHosted)
+            isValidated = await this.testPartitionKey(memberId)
+		return isValidated
+	}
+	/**
+	 * Registers a new candidate to MyLife membership
+	 * @public
+	 * @param {object} candidate { 'avatarName': string, 'email': string, 'humanName': string, }
+	 */
+	async registerCandidate(candidate){
+		return await this.#factory.registerCandidate(candidate)
+	}
+	/**
+	 * Submits and returns the memory to MyLife via API.
+	 * @public
+	 * @todo - consent check-in with spawned Member Avatar
+	 * @param {object} summary - Object with story summary and metadata
+	 * @returns {object} - The story document from Cosmos.
+	 */
+	async memory(summary){
+		if(!summary.mbr_id?.length)
+			throw new Error('story `mbr_id` required')
+		if(!summary.summary?.length)
+			throw new Error('story `summary` required')
+		summary.being = 'story'
+		summary.form = 'memory'
+		return await this.summary(summary) // @todo - convert modular
+	}
+	/**
+	 * Submits and returns a summary to MyLife via API.
+	 * @param {object} summary - Object with story summary and metadata
+	 * @returns {object} - The story document from Cosmos.
+	 */
+	async summary(summary){
+		const {
+			being='story',
+			form='story',
+			id=this.globals.newGuid,
+			mbr_id,
+			title=`untitled ${ form }`,
+		} = summary
+		const story = {
+			...summary,
+			being,
+			form,
+			id,
+			mbr_id,
+			name: `${ being }_${ title.substring(0,64) }_${ mbr_id }`,
 		}
-		return _category
+		const savedStory = this.globals.sanitize(await this.#factory.summary(story))
+		return savedStory
 	}
-	async formatQuestion(_question){
-		//	question formatting
-		return super.formatQuestion(_question)
+	/**
+	 * Tests partition key for member
+	 * @public
+	 * @param {string} _mbr_id member id
+	 * @returns {boolean} returns true if partition key is valid
+	 */
+	async testPartitionKey(_mbr_id){
+		return await this.#factory.testPartitionKey(_mbr_id)
 	}
-	//	private functions
-	async #isQuestion(_question){	//	question or statement?
-		const _model = 'curie-instruct-beta'
-		await openai.createCompletion({
-			model: _model,
-			prompt: `Is the phrase: \"${_question}\", a question (yes/no)?`,
-			temperature: 0,
-			max_tokens: 12,
-			top_p: 0.52,
-			best_of: 3,
-			frequency_penalty: 0,
-			presence_penalty: 0,
-		})
-			.then(
-				(_response)=>{
-					//	response insertion/alteration points for approval, validation, storage, pruning
-					//	challengeResponse(_response) //	insertion point: human reviewable
-					//	add relevence question
-					if(_response.data.choices[0].text.trim().toLowerCase().replace('\n','').includes('no')) _question += ', how can MyLife help?'
-				})
-		return _question
+	/* getters/setters */
+	/**
+	 * Gets MyLife agent role, refers to server entity Maht/MyLife
+	 * @returns {object} returns { role: 'system', content: 'I am <||Q||>, AI-Agent for the nonprofit member organization MyLife' }
+	*/
+	get agentRole(){
+		switch(this.being){
+			default:	//	core
+				const replacedDescription = 'I am <||Q||>, AI-Agent for the nonprofit member organization MyLife'
+				return {
+						role: "system",
+						content: replacedDescription
+					}					
+		}
 	}
+	/**
+	 * Gets being type. MyLife are _only_ MyLife prototypes, and presents as mylife, even though the current manifestation is a datacore `core` being.
+	 * @returns {string} returns 'MyLife'
+	*/
+	get being(){
+		return 'MyLife'
+	}
+	get version(){
+		return this.#version
+	}
+	set version(_version){
+		if(!this.globals.isValidVersion(_version))
+			throw new Error('Invalid version number')
+		this.#version = _version
+	}
+}
+/* exports */
+export {
+	Member,
+	Organization,
+	MyLife,
 }
