+++ inc/json-schemas/routines/about.json
@@ -21,7 +21,12 @@
         },
 		{
 			"dialog": {
-				"message": "Guiding MyLife is a board of dedicated volunteers, passionate about our mission.<ul><li>Erik Jespersen, <em>President and Treasurer</em></li><li>Kenneth Williams, <em>Vice-President</em></li><li>Stephen Kenney, <em>Technology Director</em></li><li>Russ Olivier, <em>Director</em></li></ul>In addition, we've been bolstered by numerous volunteers who have helped move our services build to completion, and currently in leadership roles, we have:<ul><li><a href=\"mailto:jeffn@humanremembranceproject.org\">Dr. Jeff Nagy</a>, Executive Director</li><li><a href=\"mailto:jessew@humanremembranceproject.org\">Jesse Waters</a>, CMO</li><li>Mark Schwanke [mark@humanremembranceproject.org], Head of Community Partnerships</li></ul>"
+				"message": "Guiding MyLife is a board of dedicated volunteers, passionate about our mission.<ul><li>Erik Jespersen, <em>President</em></li><li>Ken Williams, <em>Vice-President</em></li><li>Steve Kenney, <em>Technology Director</em></li><li>Sam Collins, <em>Director</em></li><li>Russ Olivier, <em>Director</em></li><li>Avani Patel, <em>Director</em></li></ul>"
+			}
+		},
+		{
+			"dialog": {
+				"message": "In addition, we've been bolstered by numerous volunteers who have helped move our services build to completion, and currently in volunteer leadership roles, we have:<ul><li><a href=\"mailto:jeffn@humanremembranceproject.org\">Dr. Jeff Nagy</a>, Executive Director</li><li><a href=\"mailto:jessew@humanremembranceproject.org\">Jesse Waters</a>, CMO</li><li>Mark Schwanke [mark@humanremembranceproject.org], Head of Community Partnerships</li></ul>"
 			}
 		},
 		{
