+++ views/assets/css/chat.css
@@ -6,7 +6,7 @@
 }
 /* MyLife chat container */
 /* Chat Container */
-.chat-container {
+.chat-container-old {
     display: flex;
 	flex: 1 0 60%;
     flex-direction: column;
@@ -14,6 +14,19 @@
     padding: 0.5rem;
     overflow-y: hidden; /* let the chat-section handle scrolling */
 	/* z-index: var(--chat-base-z-index); */
+}
+/* Chat Container */
+.chat-container {
+	background: #ffffffad;
+	border-radius: var(--mylife-border-radius);
+	box-shadow: 0 0 10px rgba(0, 0, 0, 0.05);
+	display: flex;
+	flex: 1 1 320px;
+	flex-direction: column;
+    max-width: 1000px; /* Prevents it from growing too large */
+	min-width: 320px; /* Prevents shrinking too much */
+	overflow-y: hidden;
+	padding: 0.5rem;
 }
 .chat-input-container {
     flex: 1 0 auto;
@@ -140,9 +153,7 @@ img.chat-message-thumb-small {
 }
 .chat-section { 
     flex: 1 1 40vh;
-    background: #ffffffad;
     border-radius: var(--mylife-border-radius);
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.05);
     display: flex;
     flex-direction: column;
     gap: 0.5rem;
