+++ views/members.html
@@ -17,7 +17,7 @@
 		<div id="audio-popup" class="audio-popup"></div>
 	</div>
 	<div id="chat-input-container" class="chat-input-container">
-		<div id="audio-icon" class="label fas fa-microphone-lines"></div>
+		<div id="audio-icon" class="fas fa-microphone-lines"></div>
 		<textarea id="chat-input-text" class="chat-input-text"></textarea>
 		<button id="chat-input-submit" class="button chat-submit" type="button" disabled>Submit</button>
 	</div>
